terraform/scim-api/lambda/code/requirements.txt
apig-wsgi==2.18.0
asgiref==3.8.1
Django==5.1
django-storages==1.14.4
djangorestframework==3.15.2
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
whitenoise==6.7.0